requirements.txt
pyyaml
chardet
requests
warcio
newspaper3k
ratelimit